fairlens-backend-v2/fairlens-backend/fairlens-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1

# Data
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
pyarrow==17.0.0

# Fairness + explainability
fairlearn==0.10.0
shap==0.46.0

# Google AI
google-generativeai==0.7.2

# PDF report
reportlab==4.2.2

# Utils
httpx==0.27.2
